+++ lib/message_port_channel.dart
@@ -1,0 +1,1 @@
+export 'src/message_port_channel.dart' show MessagePortChannel;